Brings Hypothesis API call to the template
Also changes some visual styling
