+++ styles/annotationViewer.css
@@ -1,9 +1,13 @@
-.annotation_viewer_li {
-    display: block !important;
-    margin-top: 8px;
+.obj_article_details .galleys_links li {
+    margin-right: 0.25em;
 }
 
-.annotation_viewer_span {
+.annotation_viewer {
+    margin-top: 8px;
+    visibility: hidden;
+}
+
+.annotation_viewer > span {
     padding: 4px 8px;
     border: 1px solid rgba(0,0,0,0.54);
     border-radius: 15px;
